requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
python-jose[cryptography]
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pywin32
Pillow
python-multipart
aiofiles
python-dotenv
loguru
pymupdf
jinja2
pydantic-settings
pyyaml
pytest
pytest-asyncio
httpx<0.25.0
tzdata
cairosvg
